requirements.txt
mcp
openai
pyaudio>=0.2.11
websocket-client>=1.6.0
python-dotenv>=0.19.0
aiohttp
numpy>=1.21.0
google-cloud-texttospeech>=2.16.0
tiktoken>=0.7.0
openwakeword>=0.6.0